+++ modules/configuration/src/main/scala/package.scala
@@ -1,0 +1,4 @@
+package at.ac.oeaw.imba.gerlich.gerlib
+
+/** Functionality related to configuration (e.g., from file or CLI) */
+package object configuration